gestion des erreurs lors des migrations: by koffi victorien

+++ database/migrations/2025_05_02_124135_add_documents_to_acte_naissance_table.php
@@ -13,10 +13,15 @@ return new class extends Migration
         });
     }
 
-    public function down()
-    {
+   public function down()
+{
+    if (Schema::hasTable('acte_naissance')) {
         Schema::table('acte_naissance', function (Blueprint $table) {
-            $table->dropColumn('documents');
+            if (Schema::hasColumn('acte_naissance', 'documents')) {
+                $table->dropColumn('documents');
+            }
         });
     }
+}
+
 };
